refactor : package 생성 및 분류
