+++ src/main/java/com/ann/annovation/answer/AnswerRepository.java
@@ -1,4 +1,4 @@
-package com.ann.annovation;
+package com.ann.annovation.answer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
